feat: add Provider to RichText and Annotation
