+++ src/components/RichText/Annotation/AnnotationItem.json
