+++ src/components/RichText/Annotation/AnnotationItem.stories.tsx
@@ -1,18 +1,18 @@
-import { Annotation } from "./Annotation.js";
-import RichTextAnnotationObject from "./Annotation.json";
+import { AnnotationItem } from "./AnnotationItem.js";
+import RichTextAnnotationObject from "./AnnotationItem.json";
 
 import type { Meta, StoryObj } from "@storybook/react";
 import type { RichTextItem } from "@udus/notion-libs";
 
-const meta: Meta<typeof Annotation> = {
+const meta: Meta<typeof AnnotationItem> = {
   title: "Annotation",
-  component: Annotation,
+  component: AnnotationItem,
   tags: ["autodocs"],
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Annotation>;
+type Story = StoryObj<typeof AnnotationItem>;
 
 export const Default: Story = {
   args: {
