+++ src/components/RichText/Annotation/index.ts
@@ -1,4 +1,5 @@
 export * from "./Annotation.js";
+export * from "./AnnotationItem.js";
 export * from "./Bold.js";
 export * from "./Color.js";
 export * from "./InlineCode.js";
